+++ lib/utils.zsh
@@ -20,6 +20,16 @@ spaceship::defined() {
   typeset -f + "$1" &> /dev/null
 }
 
+# Check if array includes and item
+# USAGE:
+#  spaceship::includes <array_name> <item>
+spaceship::includes() {
+  local array_name="$1" item="$2"
+  local array=("${(@P)array_name}")
+
+  (( $array[(Ie)$item] ))
+}
+
 # Precompile zsh file to ZWC (zsh word code)
 # USAGE:
 #  spaceship::precomile <file>
@@ -67,7 +77,7 @@ spaceship::is_section_async() {
   )
 
   # Some sections must be always sync
-  if (( $sync_sections[(Ie)$section] )); then
+  if spaceship::includes sync_sections "$section"; then
     return 1
   fi
 
@@ -151,9 +161,13 @@ spaceship::upsearch() {
   while [ "$root" ]; do
     # For every file as an argument
     for file in "${files[@]}"; do
-      local filepath="$(find $root -maxdepth 1 -name $file -print -quit 2>/dev/null)"
-      if [[ -n "$filepath" ]]; then
-        [[ -z "$silent" ]] && echo "$filepath"
+      local find_match="$(find $root -maxdepth 1 -name $file -print -quit 2>/dev/null)"
+      local filename="$root/$file"
+      if [[ -n "$find_match" ]]; then
+        [[ -z "$silent" ]] && echo "$find_match"
+        return 0
+      elif [[ -e "$filename" ]]; then
+        [[ -z "$silent" ]] && echo "$filename"
         return 0
       fi
     done